feat: copy buffer

+++ crates/forky_play/src/extensions/app_x.rs
@@ -10,7 +10,7 @@ pub impl App {
 		self.add_startup_system(utility::surrender_focus);
 		self
 	}
-	fn forky_exit_after(&mut self, secs: u64) -> &mut Self {
+	fn forky_exit_after(&mut self, secs: f64) -> &mut Self {
 		self.add_system(utility::create_exit_after_system(secs));
 		self
 	}
